+++ src/hooks/cart/usePatchCartItem.ts
@@ -1,5 +1,5 @@
 import { patchCartItem } from '@/api/cartItem';
-import QUERY_KEYS from '@/constants/queryKeys';
+import { QUERY_KEYS } from '@/constants/index';
 import { useToast } from '@/hooks/index';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 
@@ -26,6 +26,7 @@ const usePatchCartItem = () => {
     onError: () => {
       createToast({
         message: '⛔️ 상품을 제거하는데 실패했습니다. 다시 시도해 주세요.',
+        delayTime: 1_000 * 2,
       });
     },
   });
